+++ src/Chest/Controllers/v2/MetadataController.cs
@@ -4,24 +4,21 @@
 #pragma warning disable SA1008 // Opening parenthesis must be spaced correctly
 #pragma warning disable SA1300 // Element must begin with upper-case letter
 
-namespace Chest.Controllers
+namespace Chest.Controllers.v2
 {
-    using System;
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.Linq;
     using System.Net;
     using System.Threading.Tasks;
     using Chest.Exceptions;
-    using Chest.Models.v1;
+    using Chest.Models.v2;
     using Chest.Services;
     using Microsoft.AspNetCore.Mvc;
-    using Newtonsoft.Json;
     using Swashbuckle.AspNetCore.SwaggerGen;
 
-    [Obsolete("MetadataController is obsolete, please use v2/MetadataController instead.")]
-    [ApiVersion("1")]
-    [Route("api")]
+    [ApiVersion("2")]
+    [Route("api/v{version:apiVersion}/")]
     public class MetadataController : ControllerBase
     {
         private readonly IDataService service;
@@ -45,16 +42,14 @@ namespace Chest.Controllers
 
             try
             {
-                var serializedData = JsonConvert.SerializeObject(model.Data);
-                var serializedKeywords = model.Keywords == null ? string.Empty : JsonConvert.SerializeObject(model.Keywords);
-                await this.service.Add(category, collection, key, serializedData, serializedKeywords);
+                await this.service.Add(category, collection, key, model.Data, model.Keywords);
             }
             catch (DuplicateKeyException)
             {
                 return this.StatusCode((int)HttpStatusCode.Conflict, new { Message = $"Data already exists for category: {category} collection: {collection} key: {key}" });
             }
 
-            return this.Created(this.Request.GetRelativeUrl($"api/{category}/{collection}/{key}"), model);
+            return this.Created(this.Request.GetRelativeUrl($"api/v2/{category}/{collection}/{key}"), model);
         }
 
         [HttpPost("{category}/{collection}")]
@@ -71,13 +66,7 @@ namespace Chest.Controllers
 
             try
             {
-                var serializedModel = model.ToDictionary(x => x.Key, x =>
-                {
-                    var serializedData = JsonConvert.SerializeObject(x.Value.Data);
-                    var serializedKeywords = x.Value.Keywords == null ? string.Empty : JsonConvert.SerializeObject(x.Value.Keywords);
-                    return (serializedData, serializedKeywords);
-                });
-                await this.service.BulkAdd(category, collection, serializedModel);
+                await this.service.BulkAdd(category, collection, model.ToDictionary(x => x.Key, x => (x.Value.Data, x.Value.Keywords)));
             }
             catch (DuplicateKeyException e)
             {
@@ -102,9 +91,7 @@ namespace Chest.Controllers
 
             try
             {
-                var serializedData = JsonConvert.SerializeObject(model.Data);
-                var serializedKeywords = model.Keywords == null ? string.Empty : JsonConvert.SerializeObject(model.Keywords);
-                await this.service.Update(category, collection, key, serializedData, serializedKeywords);
+                await this.service.Update(category, collection, key, model.Data, model.Keywords);
             }
             catch (NotFoundException)
             {
@@ -128,13 +115,7 @@ namespace Chest.Controllers
 
             try
             {
-                var serializedModel = model.ToDictionary(x => x.Key, x =>
-                {
-                    var serializedData = JsonConvert.SerializeObject(x.Value.Data);
-                    var serializedKeywords = x.Value.Keywords == null ? string.Empty : JsonConvert.SerializeObject(x.Value.Keywords);
-                    return (serializedData, serializedKeywords);
-                });
-                await this.service.BulkUpdate(category, collection, serializedModel);
+                await this.service.BulkUpdate(category, collection, model.ToDictionary(x => x.Key, x => (x.Value.Data, x.Value.Keywords)));
             }
             catch (NotFoundException e)
             {
@@ -192,7 +173,7 @@ namespace Chest.Controllers
 
         [HttpGet("{category}/{collection}")]
         [SwaggerOperation("Metadata_GetKeysWithData")]
-        [SwaggerResponse((int)HttpStatusCode.OK, typeof(Dictionary<string, Dictionary<string, string>>))]
+        [SwaggerResponse((int)HttpStatusCode.OK, typeof(Dictionary<string, string>))]
         [SwaggerResponse((int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetKeysWithData(string category, string collection, [FromQuery]string keyword)
         {
@@ -203,16 +184,14 @@ namespace Chest.Controllers
                 return this.NotFound(new { Message = $"No record found for Category: {category} Collection: {collection} filtered by Keyword: {keyword}" });
             }
 
-            var deserializedKeyValueData = keyValueData.Select(keyValue => new Dictionary<string, Dictionary<string, string>> { { keyValue.Key, JsonConvert.DeserializeObject<Dictionary<string, string>>(keyValue.Value) } });
-
-            return this.Ok(deserializedKeyValueData);
+            return this.Ok(keyValueData);
         }
 
         // NOTE: This is POST because passing around massive strings in a query parameter might
         // hit some URL length limitation along the way
         [HttpPost("{category}/{collection}/find")]
         [SwaggerOperation("Metadata_FindByKeys")]
-        [SwaggerResponse((int)HttpStatusCode.OK, typeof(IDictionary<string, IDictionary<string, string>>))]
+        [SwaggerResponse((int)HttpStatusCode.OK, typeof(IDictionary<string, string>))]
         [SwaggerResponse((int)HttpStatusCode.NotFound)]
         [SwaggerResponse((int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> FindByKeys(
@@ -251,10 +230,7 @@ namespace Chest.Controllers
                 return this.NotFound(new { Message = $"No data found for category: {category} collection: {collection} and key: {key}" });
             }
 
-            var deserializedData = JsonConvert.DeserializeObject<Dictionary<string, string>>(data);
-            var deserializedKeywords = string.IsNullOrWhiteSpace(keywords) ? default(List<string>) : JsonConvert.DeserializeObject<List<string>>(keywords);
-
-            return this.Ok(new MetadataModel { Data = deserializedData, Keywords = deserializedKeywords });
+            return this.Ok(new MetadataModel { Data = data, Keywords = keywords });
         }
     }
 }